requirements.txt
pdfminer.six
python-docx
chardet
spacy
scikit-learn

